Refactor: Lazy load on PostGrid issue (#864)
<!--
  Thanks for contributing!

Provide a description of your changes below and a general summary in the
title

Please look at the following checklist to ensure that your PR can be
accepted quickly:
-->

## Description
This PR introduces a refactor in order to fix and enhance the current
issue with lazy loading, specifically with the `PostGrid` block. Related
to the following issue: https://github.com/flutter/flutter/issues/119136

The main solution involves refactoring the `PostGrid` to become a
SliverGrid, allowing for performant scrolling with a big number of
items. Since the PostGrid layouts its first children as a `PostLarge`
and the rest of them as a `PostMedium`, it was needed to create a custom
`gridDelegate` that allowed the grid to handle its children in the
mentioned way.

Since now the `PostGrid` widget is built as a sliver, this also involved
a refactor on `CategoryFeedItem`, so that every block widget could be
built as a sliver. `CategoryFeed` widget was also refactored, so that
the feed content list is built inside a `CustomScrollView`.

As a consequence of having `CategoryFeedItem` build every block as a
sliver, the screens that made use of this widget were also refactored
(`Article screen` and `Search screen`), resulting in performant
scrolling on this screens aswell.


Example video showing the feed with a grid that holds more than 1000
items:


https://user-images.githubusercontent.com/48563141/236327806-fa8f5c6c-f9c8-4c2b-9ab5-db108904cc92.mp4

Example video showing the article screen:


https://user-images.githubusercontent.com/48563141/236330277-bac7c73d-4832-4a10-83ad-e7859c99a311.mp4



<!--- Describe your changes in detail -->

## Type of Change

<!--- Put an `x` in all the boxes that apply: -->

- [ ] ✨ New feature (non-breaking change which adds functionality)
- [ ] 🛠️ Bug fix (non-breaking change which fixes an issue)
- [ ] ❌ Breaking change (fix or feature that would cause existing
functionality to change)
- [x] 🧹 Code refactor
- [ ] ✅ Build configuration change
- [ ] 📝 Documentation
- [ ] 🗑️ Chore

---------

Co-authored-by: Tom Arra <tarra3@gmail.com>