Fix: model json serializable generation (#1055)
<!--
  Thanks for contributing!

Provide a description of your changes below and a general summary in the
title

Please look at the following checklist to ensure that your PR can be
accepted quickly:
-->

## Description
This is related to the following issue reported on the Flutter
repository https://github.com/flutter/flutter/issues/122899

When running `build_runner` to generate the generated code of the models
from the api, it was causing issues with some of the api tests.

- Changed return type of `NewsBlockConverter()` from List<Map<String,
dynamic>> `List<dynamic>`. This was causing tests to fail.
- Added `explicitToJson` statements to some models that had complex
attributes inside, like another model. This was also causing tests to
fail.
- Since generated files are pushed to the repo, updated the .g files
with the new generated ones.


<!--- Describe your changes in detail -->



## Type of Change

<!--- Put an `x` in all the boxes that apply: -->

- [ ] ✨ New feature (non-breaking change which adds functionality)
- [x] 🛠️ Bug fix (non-breaking change which fixes an issue)
- [ ] ❌ Breaking change (fix or feature that would cause existing
functionality to change)
- [ ] 🧹 Code refactor
- [ ] ✅ Build configuration change
- [ ] 📝 Documentation
- [ ] 🗑️ Chore