Chore: v1.0.7 (#1069)
<!--
  Thanks for contributing!

Provide a description of your changes below and a general summary in the
title

Please look at the following checklist to ensure that your PR can be
accepted quickly:
-->

## Description
- Fixed issue when running `build_runner` to generate the generated code
for the API models, since it was causing some of the API tests to fail.

- Changed return type of `NewsBlockConverter()` from List<Map<String,
dynamic>> `List<dynamic>`. This was causing tests to fail.
- Added `explicitToJson` statements to some models that had complex
attributes inside, like another model. This was also causing tests to
fail.

- Upgraded `brick.yaml` version and added new changes to `changelog.md`
file.
<!--- Describe your changes in detail -->

## Type of Change

<!--- Put an `x` in all the boxes that apply: -->

- [ ] ✨ New feature (non-breaking change which adds functionality)
- [x] 🛠️ Bug fix (non-breaking change which fixes an issue)
- [ ] ❌ Breaking change (fix or feature that would cause existing
functionality to change)
- [ ] 🧹 Code refactor
- [ ] ✅ Build configuration change
- [ ] 📝 Documentation
- [x] 🗑️ Chore